fix(block exchange): decrements block retries count when downloading with no peers (#1343)
There is a limitation in `downloadInternal` that infinitely loops for peers when attempting to download a block if the node does not have peers. Currently, the block retry counter is only decremented if there exists peers and a download attempt is made, otherwise the loop will continue until peers are found, ad infinium. This fix decrements the retry counter on each loop that no peers are found. Additionally, it waits for the `DiscoveryRateLimit` interval before retrying, as otherwise looking for peers would not produce new results.